+++ structurizr-core/src/com/structurizr/view/LineStyle.java
@@ -1,0 +1,9 @@
+package com.structurizr.view;
+
+public enum LineStyle {
+
+    Dashed,
+    Dotted,
+    Solid
+
+}